Share getTokenServiceEnv between the admin and the envelope S3 config
