+++ shared/seismic/envelopes/envelope-config.ts
@@ -62,12 +62,13 @@ function getUrlParam(name: string): string | null {
 export const S3_BUCKET = "models-resources";
 export const S3_STAGING_BUCKET = "models-resources-qa";
 
-function getS3Bucket() {
-  if (getUrlParam("tokenServiceEnv") === "staging") {
-    return S3_STAGING_BUCKET;
-  }
+/** Token-service environment for envelope uploads; ?tokenServiceEnv=staging for testing. */
+export function getTokenServiceEnv(): "staging" | "production" {
+  return getUrlParam("tokenServiceEnv") === "staging" ? "staging" : "production";
+}
 
-  return S3_BUCKET;
+function getS3Bucket() {
+  return getTokenServiceEnv() === "staging" ? S3_STAGING_BUCKET : S3_BUCKET;
 }
 
 /** AWS region of the envelope tile bucket. */
