+++ src/seismic-admin/components/app.tsx
@@ -3,7 +3,8 @@ import React, { useEffect, useState } from "react";
 import { SeismicAdminStoreContext } from "../hooks/use-seismic-admin-stores";
 import { initAdminFirebase } from "../utils/admin-firebase";
 import { loadCatalog } from "../utils/load-catalog";
-import { consumeAccessTokenFromLocation, getTokenServiceEnv, makePortalJwtGetter } from "../utils/portal-auth";
+import { getTokenServiceEnv } from "../../../shared/seismic/envelopes/envelope-config";
+import { consumeAccessTokenFromLocation, makePortalJwtGetter } from "../utils/portal-auth";
 import { SeismicAdminStore } from "../seismic-admin-store";
 import { AdminBody } from "./admin-body";
 import { AdminHeader } from "./admin-header";
